+++ packages/components/.npmignore
@@ -22,5 +22,6 @@ tsconfig.test.json
 # Source code
 src/components/*/*.stories.tsx
 src/components/*/*.test.tsx
+src/utils/hooks/useWebStorybookColorScheme.ts
 src/utils/storybook.tsx
 src/App.tsx
